+++ changelog.md
@@ -1,3 +1,10 @@
+## 3.2.7
+
+- Auto update of the elevation parameter of the rider when mount and dismount some specific mount
+- New active effect manager for add and remove custom active effect when mount and dismount
+- Add token magic effect for apply the "flying" effect
+- Set up for the new token attacher paramters for the mount up macro "Add canMoveConstrained flag and API to set this flag so a attached element is allowed to move within area of base token. This is only supported for Tokens for now."
+
 ## 3.2.5 
 
 - Removed old flag set